requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-firestore
langchain-google-vertexai